3D Recounstraction of Micro Parts Using LabView
the scanner is capable of reconstructing the surface topography of a small object that has a maximum height of less than 30 mm. The 3D surface reconstruction is then treated using digital signal processing techniques to output a better representation of the actual object.

A laser displacement sensor is used to take a series of elevation data across an object. The data then used to reconstruct the 3d surface of an object. X-Y stage used to move the object so that the laser could take measurements at a different cross-section of the object. The data then possessed and analyzed using MATLAB.